mediateka 19 commit 16 sprint
