+++ app/src/main/java/com/practicum/playlismaker/base/data/NetworkManager.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlismaker.base.data
+
+import com.practicum.playlismaker.base.data.dto.BaseResponse
+
+interface NetworkManager {
+    fun doRequest(dto: Any): BaseResponse
+}
