+++ app/src/main/java/com/practicum/playlismaker/base/data/dto/BaseResponse.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlismaker.base.data.dto
+
+open class BaseResponse {
+    var resultCode = 0
+}
