--- app/src/main/java/com/practicum/playlismaker/data/NetworkManager.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlismaker.data
-
-import com.practicum.playlismaker.data.dto.BaseResponse
-
-interface NetworkManager {
-    fun doRequest(dto: Any): BaseResponse
-}
