+++ app/src/main/java/com/practicum/playlismaker/data/TracksRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package com.practicum.playlismaker.data
 
+import com.practicum.playlismaker.base.data.NetworkManager
 import com.practicum.playlismaker.data.dto.ITunesResponseDto
 import com.practicum.playlismaker.data.dto.TracksSearchRequest
-import com.practicum.playlismaker.domain.api.TracksRepository
-import com.practicum.playlismaker.domain.models.Track
-import com.practicum.playlismaker.ui.formatTime
+import com.practicum.playlismaker.search.domain.api.TracksRepository
+import com.practicum.playlismaker.search.domain.models.Track
 
 class TracksRepositoryImpl(private val networkManager: NetworkManager) : TracksRepository {
