--- app/src/main/java/com/practicum/playlismaker/data/dto/BaseResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlismaker.data.dto
-
-open class BaseResponse {
-    var resultCode = 0
-}
