+++ app/src/main/java/com/practicum/playlismaker/data/dto/ITunesResponseDto.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlismaker.data.dto
 
-import com.practicum.playlismaker.domain.models.Track
+import com.practicum.playlismaker.base.data.dto.BaseResponse
 
 data class ITunesResponseDto(
     val resultCount: Int, val results: List<TrackDto>
