--- app/src/main/java/com/practicum/playlismaker/domain/api/SearchHistoryInteractor.kt
@@ -1,10 +1,0 @@
-package com.practicum.playlismaker.domain.api
-
-import com.practicum.playlismaker.domain.models.Track
-
-interface SearchHistoryInteractor {
-
-    fun update(track: Track)
-    fun getHistory(): MutableList<Track>
-    fun clear()
-}
