--- app/src/main/java/com/practicum/playlismaker/domain/api/SearchHistoryRepository.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlismaker.domain.api
-
-import com.practicum.playlismaker.domain.models.Track
-
-interface SearchHistoryRepository {
-    fun update(track: Track)
-    fun getHistory(): MutableList<Track>
-    fun clear()
-}
