--- app/src/main/java/com/practicum/playlismaker/domain/api/TracksRepository.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlismaker.domain.api
-
-import com.practicum.playlismaker.domain.models.Track
-
-interface TracksRepository {
-    fun searchTracks(search: String): List<Track>
-}
