+++ app/src/main/java/com/practicum/playlismaker/search/domain/api/SearchHistoryInteractor.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlismaker.search.domain.api
+
+import com.practicum.playlismaker.search.domain.models.Track
+
+interface SearchHistoryInteractor {
+
+    fun update(track: Track)
+    fun getHistory(): MutableList<Track>
+    fun clear()
+}
