+++ app/src/main/java/com/practicum/playlismaker/search/domain/api/SearchHistoryRepository.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlismaker.search.domain.api
+
+import com.practicum.playlismaker.search.domain.models.Track
+
+interface SearchHistoryRepository {
+    fun update(track: Track)
+    fun getHistory(): MutableList<Track>
+    fun clear()
+}
