+++ app/src/main/java/com/practicum/playlismaker/search/domain/api/TracksRepository.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlismaker.search.domain.api
+
+import com.practicum.playlismaker.search.domain.models.Track
+
+interface TracksRepository {
+    fun searchTracks(search: String): List<Track>
+}
