+++ app/src/main/java/com/practicum/playlismaker/search/domain/impl/TracksInteractorImpl.kt
@@ -1,0 +1,15 @@
+package com.practicum.playlismaker.search.domain.impl
+
+import com.practicum.playlismaker.search.domain.api.TracksInteractor
+import com.practicum.playlismaker.search.domain.api.TracksRepository
+import java.util.concurrent.Executors
+
+class TracksInteractorImpl(private val repository: TracksRepository) : TracksInteractor {
+
+    private val executor = Executors.newCachedThreadPool()
+    override fun searchTracks(search: String, consumer: TracksInteractor.TracksConsumer) {
+        executor.execute {
+            consumer.consume(repository.searchTracks(search))
+        }
+    }
+}
