--- app/src/main/java/com/practicum/playlismaker/ui/MediatekaActivity.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlismaker.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.practicum.playlismaker.R
-
-class MediatekaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mediateka)
-    }
-}
