+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:background="@color/background_color"
     android:orientation="vertical"
     android:padding="@dimen/padding"
-    tools:context=".ui.MainActivity">
+    tools:context=".main.ui.MainActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:fontFamily = "@font/ys_display_medium"
