+++ app/src/main/res/layout/activity_player.xml
@@ -26,7 +26,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:paddingBottom="@dimen/player_margin_common"
-            tools:context=".ui.player.PlayerActivity">
+            tools:context=".player.ui.PlayerActivity">
 
             <ImageView
                 android:id="@+id/player_artwork"
