+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/background_black_white"
     android:orientation="vertical"
-    tools:context=".ui.search.SearchActivity">
+    tools:context=".search.ui.SearchActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
